requirements.txt
# Core Memorix SDK
memorix-sdk==0.1.0

# Vector Stores
faiss-cpu==1.7.4
qdrant-client==1.6.0

# Embedding Models
openai==1.3.0
google-generativeai==0.3.0
sentence-transformers==2.2.0

# Web Frameworks
streamlit==1.28.0
gradio==4.0.0
flask==2.3.0
fastapi==0.104.0
uvicorn==0.24.0

# Data Processing
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0

# Document Processing
pypdf==3.15.0
python-docx==0.8.11
markdown==3.5.0

# Visualization
matplotlib==3.7.0
seaborn==0.12.0
plotly==5.17.0

# Jupyter
jupyter==1.0.0
ipywidgets==8.0.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0
tqdm==4.66.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.0.0
flake8==6.0.0
mypy==1.5.0

# Optional: GPU support
# faiss-gpu==1.7.4
# torch==2.0.0
# torchvision==0.15.0 